fixed up tests
